Move post fetching to group index route
